A lot has been done.
Balltracker keeps track (surprisingly enough) off all balls that have been linked so far and manages score calculation, it has a field that is checked by the GameView which exposes that it is ready to update the score. It is bad practice to have the View handle this, we can look into that later on.